refactor: ajustes no footer e navbar
