+++ frontend/src/App.js
@@ -2,9 +2,9 @@ import { Box, Container } from '@mui/material';
 import './App.css';
 import CadastroCliente from './paginas/CadastroCliente/CadastroCliente';
 import EditarCliente from './paginas/EditarCliente/EditarCliente';
-import Footer from './components/Footer';
+import Footer from './componentes/Footer';
 import GerenciarClientes from './paginas/GerenciarClientes/GerenciarClientes';
-import Navbar from './components/Navbar';
+import Navbar from './componentes/Navbar';
 
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
