+++ frontend/src/componentes/Footer.js
@@ -2,7 +2,7 @@ import { Box, Typography } from "@mui/material";
 
 export default function Footer() {
     return (
-        <Box component="footer" className='Footer' padding={4}>
+        <Box component="footer" className='Footer' padding={4} borderTop={"solid 1px #AAAAAA"}>
             <Typography>Comércios S.A.</Typography>
         </Box>
     );
