+++ frontend/src/componentes/Navbar.js
@@ -4,7 +4,7 @@ import HomeIcon from '@mui/icons-material/Home';
 
 export default function Navbar() {
   return (
-    <Box component="nav" className='Navbar' sx={{ padding: 2, gap:8, display:"flex" }}>
+    <Box component="nav" className='Navbar' sx={{ padding: 2, gap:8, display:"flex", borderBottom: "solid 1px #AAAAAA" }}>
       <Link href="/" sx={{ display: "flex", textDecoration: "none", alignItems: "center" }}>
         <HomeIcon fontSize="large" />
         <Typography fontSize="large">Home</Typography>
